Factor out progress-util.ts.
